mcp-server/requirements.txt
annotated-types          0.7.0
anyio                    4.9.0
cachetools               5.5.2
certifi                  2025.4.26
charset-normalizer       3.4.2
click                    8.1.8
google-api-core          2.24.2
google-api-python-client 2.169.0
google-auth              2.40.1
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.2
googleapis-common-protos 1.70.0
h11                      0.16.0
httpcore                 1.0.9
httplib2                 0.22.0
httpx                    0.28.1
httpx-sse                0.4.0
idna                     3.10
mcp                      1.8.0
oauthlib                 3.2.2
pip                      25.1.1
proto-plus               1.26.1
protobuf                 6.30.2
pyasn1                   0.6.1
pyasn1_modules           0.4.2
pydantic                 2.11.4
pydantic_core            2.33.2
pydantic-settings        2.9.1
pyparsing                3.2.3
python-dotenv            1.1.0
python-multipart         0.0.20
requests                 2.32.3
requests-oauthlib        2.0.0
rsa                      4.9.1
sniffio                  1.3.1
sse-starlette            2.3.4
starlette                0.46.2
typing_extensions        4.13.2
typing-inspection        0.4.0
uritemplate              4.1.1
urllib3                  2.4.0
uvicorn                  0.34.2